requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
# flask_sockets
# gevent
groq
openai
cohere
python-dotenv
flask-cors
pinecone
# transformers
# torch
# langchain
# pypdf
# sentence-transformers
# gunicorn
pdf2image